Confirmed new makefile template works with Chapter3 examples and exercises up to Ex3_17. Reviewed code in the implementations

+++ Chapter3/Arrays/Exercises/Ex3_11/ex3_11.c
@@ -19,7 +19,7 @@ Program 3.11
  */
 int main(int argc, char* argv[argc + 1]) {
 
-    int a[99];
+    size_t a[99];
     for (size_t i = 0; i < 99; i++) a[i] = 98 - i;
     for (size_t i = 0; i < 99; i++) a[i] = a[a[i]];
 
@@ -29,7 +29,7 @@ int main(int argc, char* argv[argc + 1]) {
     printf("\n");
 
     for (size_t i = 0; i < 99; i++) {
-        printf(" %d ", a[i]);
+        printf(" %zu ", a[i]);
     }
     printf("\n");
     return EXIT_SUCCESS;
